// src/projects/enums/project-messages.enum.ts
export enum ProjectSuccessMessages {
  CREATED = "پروژه ایجاد شد",
  FOUND = "پروژه پیدا شد",
  LISTED = "لیست پروژه ها دریافت شد",
  UPDATED = "پروژه با موفقیت بروز رسانی شد",
  DELETED = "پروژه با موفقیت حذف شد",
}

export enum ProjectErrorMessages {
  DUPLICATE = "پروژه تکراری است",
  NOT_FOUND = "پروژه پیدا نشد",
}